requirements_prod.txt
-r requirements.txt

# Production Dependencies
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
sentry-sdk==1.39.1
django-storages==1.14.2
boto3==1.34.14
whitenoise==6.6.0
django-cors-headers==4.3.1

# Security
django-axes==6.3.0
django-csp==3.7
django-ratelimit==4.1.0 